requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Gemini 3 SDK (Unified 2026 Version)
google-genai>=1.60.0
google-api-core>=2.20.0
# Security and Data
pydantic>=2.10.0
python-multipart>=0.0.10

# Performance (Optional but recommended for high-traffic APIs)
httpx[socks]>=0.28.0

# Database
psycopg2-binary>=2.9.0

# Vector operations
numpy>=1.24.0

# Environment variables
python-dotenv>=1.0.0